requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
python-dotenv==1.0.1
psycopg==3.2.3
